requirements.txt
streamlit==1.16.0
streamlit-chat==0.0.2.1
requests==2.28.1
openai==0.25.0
python-dotenv==0.21.0
SQLAlchemy==1.4.46
python-jose==3.3.0
passlib==1.7.4
fastapi==0.88.0
bcrypt==4.0.1
cryptography==39.0.0
psycopg2==2.9.5
uvicorn==0.20.0